requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.7
django-filter==2.4.0
django-heroku==0.3.1
djangorestframework==3.12.2
Markdown==3.3.4
psycopg2==2.8.6
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
gunicorn==20.4.0
